test: Split unittests for documents
